+++ python/cusignal/__init__.py
@@ -108,6 +108,9 @@ from cusignal.utils.arraytools import (
     get_shared_array,
     get_shared_mem
 )
+from cusignal._precompile import (
+    precompile_kernels
+)
 
 # Versioneer
 from ._version import get_versions
