+++ python/cusignal/_spectral.py
@@ -12,198 +12,19 @@
 # limitations under the License.
 
 import cupy as cp
-import itertools
-import numpy as np
 import warnings
 
-from enum import Enum
-from math import sin, cos, atan2
-from numba import cuda, float64, void
-from string import Template
-
-from ._precompile import _stream_cupy_to_numba
+from _caches import _cupy_kernel_cache, _numba_kernel_cache
+from _precompile import (
+    _stream_cupy_to_numba,
+    # _cupy_kernel_cache,
+    # _numba_kernel_cache,
+    _populate_kernel_cache,
+    GPUKernel,
+)
 
 # Display FutureWarnings only once per module
 warnings.simplefilter("once", FutureWarning)
-
-
-class GPUKernel(Enum):
-    LOMBSCARGLE = 0
-
-
-class GPUBackend(Enum):
-    CUPY = 0
-    NUMBA = 1
-
-
-# Numba type supported and corresponding C type
-_SUPPORTED_TYPES = {
-    np.float64: [float64, "double"],
-}
-
-
-_numba_kernel_cache = {}
-_cupy_kernel_cache = {}
-
-
-def _numba_lombscargle(x, y, freqs, pgram, y_dot):
-    """
-    _lombscargle(x, y, freqs)
-    Computes the Lomb-Scargle periodogram.
-    Parameters
-    ----------
-    x : array_like
-        Sample times.
-    y : array_like
-        Measurement values (must be registered so the mean is zero).
-    freqs : array_like
-        Angular frequencies for output periodogram.
-    Returns
-    -------
-    pgram : array_like
-        Lomb-Scargle periodogram.
-    Raises
-    ------
-    ValueError
-        If the input arrays `x` and `y` do not have the same shape.
-    See also
-    --------
-    lombscargle
-    """
-
-    F = cuda.grid(1)
-    strideF = cuda.gridsize(1)
-
-    if not y_dot[0]:
-        yD = 1.0
-    else:
-        yD = 2.0 / y_dot[0]
-
-    for i in range(F, freqs.shape[0], strideF):
-
-        # Copy data to registers
-        freq = freqs[i]
-
-        xc = 0.0
-        xs = 0.0
-        cc = 0.0
-        ss = 0.0
-        cs = 0.0
-
-        for j in range(x.shape[0]):
-
-            c = cos(freq * x[j])
-            s = sin(freq * x[j])
-
-            xc += y[j] * c
-            xs += y[j] * s
-            cc += c * c
-            ss += s * s
-            cs += c * s
-
-        tau = atan2(2.0 * cs, cc - ss) / (2.0 * freq)
-        c_tau = cos(freq * tau)
-        s_tau = sin(freq * tau)
-        c_tau2 = c_tau * c_tau
-        s_tau2 = s_tau * s_tau
-        cs_tau = 2.0 * c_tau * s_tau
-
-        pgram[i] = (
-            0.5
-            * (
-                (
-                    (c_tau * xc + s_tau * xs) ** 2
-                    / (c_tau2 * cc + cs_tau * cs + s_tau2 * ss)
-                )
-                + (
-                    (c_tau * xs - s_tau * xc) ** 2
-                    / (c_tau2 * ss - cs_tau * cs + s_tau2 * cc)
-                )
-            )
-        ) * yD
-
-
-def _numba_lombscargle_signature(ty):
-    return void(
-        ty[:], ty[:], ty[:], ty[:], ty[:],  # x  # y  # freqs  # pgram  # y_dot
-    )
-
-
-# Custom Cupy raw kernel implementing lombscargle operation
-# Matthew Nicely - mnicely@nvidia.com
-loaded_from_source = Template(
-    """
-extern "C" {
-    __global__ void _cupy_lombscargle(
-            const int x_shape,
-            const int freqs_shape,
-            const ${datatype} * __restrict__ x,
-            const ${datatype} * __restrict__ y,
-            const ${datatype} * __restrict__ freqs,
-            ${datatype} * __restrict__ pgram,
-            const ${datatype} * __restrict__ y_dot
-            ) {
-
-        const int tx {
-            static_cast<int>( blockIdx.x * blockDim.x + threadIdx.x ) };
-        const int stride { static_cast<int>( blockDim.x * gridDim.x ) };
-
-        ${datatype} yD {};
-        if ( y_dot[0] == 0 ) {
-            yD = 1.0f;
-        } else {
-            yD = 2.0f / y_dot[0];
-        }
-
-        for ( int tid = tx; tid < freqs_shape; tid += stride ) {
-
-            ${datatype} freq { freqs[tid] };
-
-            ${datatype} xc {};
-            ${datatype} xs {};
-            ${datatype} cc {};
-            ${datatype} ss {};
-            ${datatype} cs {};
-            ${datatype} c {};
-            ${datatype} s {};
-
-            for ( int j = 0; j < x_shape; j++ ) {
-                c = cos( freq * x[j] );
-                s = sin( freq * x[j] );
-
-                xc += y[j] * c;
-                xs += y[j] * s;
-                cc += c * c;
-                ss += s * s;
-                cs += c * s;
-            }
-
-            ${datatype} tau { atan2( 2.0f * cs, cc - ss ) / ( 2.0f * freq ) };
-            ${datatype} c_tau { cos(freq * tau) };
-            ${datatype} s_tau { sin(freq * tau) };
-            ${datatype} c_tau2 { c_tau * c_tau };
-            ${datatype} s_tau2 { s_tau * s_tau };
-            ${datatype} cs_tau { 2.0f * c_tau * s_tau };
-
-            pgram[tid] = (
-                0.5f * (
-                   (
-                       ( c_tau * xc + s_tau * xs )
-                       * ( c_tau * xc + s_tau * xs )
-                       / ( c_tau2 * cc + cs_tau * cs + s_tau2 * ss )
-                    )
-                   + (
-                       ( c_tau * xs - s_tau * xc )
-                       * ( c_tau * xs - s_tau * xc )
-                       / ( c_tau2 * ss - cs_tau * cs + s_tau2 * cc )
-                    )
-                )
-            ) * yD;
-        }
-    }
-}
-"""
-)
 
 
 class _cupy_lombscargle_wrapper(object):
@@ -239,7 +60,7 @@ class _cupy_lombscargle_wrapper(object):
 def _get_backend_kernel(dtype, grid, block, stream, use_numba, k_type):
 
     if not use_numba:
-        kernel = _cupy_kernel_cache[(dtype.name, k_type)]
+        kernel = _cupy_kernel_cache[(dtype.name, k_type.value)]
         if kernel:
             return _cupy_lombscargle_wrapper(grid, block, stream, kernel)
         else:
@@ -255,7 +76,7 @@ def _get_backend_kernel(dtype, grid, block, stream, use_numba, k_type):
         )
 
         nb_stream = _stream_cupy_to_numba(stream)
-        kernel = _numba_kernel_cache[(dtype.name, k_type)]
+        kernel = _numba_kernel_cache[(dtype.name, k_type.value)]
 
         if kernel:
             return kernel[grid, block, nb_stream]
@@ -267,130 +88,6 @@ def _get_backend_kernel(dtype, grid, block, stream, use_numba, k_type):
     raise NotImplementedError(
         "No kernel found for datatype {}".format(dtype.name)
     )
-
-
-def _populate_kernel_cache(np_type, use_numba, k_type):
-
-    # Check in np_type is a supported option
-    try:
-        numba_type, c_type = _SUPPORTED_TYPES[np_type]
-
-    except ValueError:
-        raise Exception("No kernel found for datatype {}".format(np_type))
-
-    # Check if use_numba is support
-    try:
-        GPUBackend(use_numba)
-
-    except ValueError:
-        raise
-
-    # Check if use_numba is support
-    try:
-        GPUKernel(k_type)
-
-    except ValueError:
-        raise
-
-    # JIT compile the numba kernels
-    if not use_numba:
-        if (str(numba_type), k_type) in _cupy_kernel_cache:
-            return
-        # Instantiate the cupy kernel for this type and compile
-        src = loaded_from_source.substitute(datatype=c_type)
-        module = cp.RawModule(
-            code=src, options=("-std=c++11", "-use_fast_math")
-        )
-        if k_type == GPUKernel.LOMBSCARGLE:
-            _cupy_kernel_cache[
-                (str(numba_type), k_type)
-            ] = module.get_function("_cupy_lombscargle")
-        else:
-            raise NotImplementedError(
-                "No kernel found for k_type {}, datatype {}".format(
-                    k_type, str(numba_type)
-                )
-            )
-
-    else:
-        if (str(numba_type), k_type) in _numba_kernel_cache:
-            return
-        # JIT compile the numba kernels
-        sig = _numba_lombscargle_signature(numba_type)
-        if k_type == GPUKernel.LOMBSCARGLE:
-            _numba_kernel_cache[(str(numba_type), k_type)] = cuda.jit(
-                sig, fastmath=True
-            )(_numba_lombscargle)
-        else:
-            raise NotImplementedError(
-                "No kernel found for k_type {}, datatype {}".format(
-                    k_type, str(numba_type)
-                )
-            )
-
-
-def precompile_kernels(dtype=None, backend=None, k_type=None):
-    r"""
-    Precompile GPU kernels for later use.
-
-    Parameters
-    ----------
-    dtype : numpy datatype or list of datatypes, optional
-        Data types for which kernels should be precompiled. If not
-        specified, all supported data types will be precompiled.
-        Specific to this unit
-            np.float64
-    backend : GPUBackend, optional
-        Which GPU backend to precompile for. If not specified,
-        all supported backends will be precompiled.
-        Specific to this unit
-            GPUBackend.CUPY
-            GPUBackend.NUMBA
-    k_type : GPUKernel, optional
-        Which GPU kernel to compile for. If not specified,
-        all supported kernels will be precompiled.
-        Specific to this unit
-            GPUBackend.LOMBSCARGLE
-        Examples
-    ----------
-    To precompile all kernels in this unit
-    >>> import cusignal
-    >>> from cusignal._upfirdn import GPUBackend, GPUKernel
-    >>> cusignal._spectral.precompile_kernels()
-
-    To precompile a specific NumPy datatype, CuPy backend, and kernel type
-    >>> cusignal._spectral.precompile_kernels( [np.float64],
-        [GPUBackend.CUPY],
-        [GPUKernel.LOMBSCARGLE],)
-
-
-    To precompile a specific NumPy datatype and kernel type,
-    but both Numba and CuPY variations
-    >>> cusignal._spectral.precompile_kernels( dtype=[np.float64],
-        k_type=[GPUKernel.LOMBSCARGLE],)
-    """
-    if dtype is not None and not hasattr(dtype, "__iter__"):
-        raise TypeError(
-            "dtype ({}) should be in list - e.g [np.float32,]".format(dtype)
-        )
-
-    elif backend is not None and not hasattr(backend, "__iter__"):
-        raise TypeError(
-            "backend ({}) should be in list - e.g [{},]".format(
-                backend, backend
-            )
-        )
-    elif k_type is not None and not hasattr(k_type, "__iter__"):
-        raise TypeError(
-            "k_type ({}) should be in list - e.g [{},]".format(k_type, k_type)
-        )
-    else:
-        dtype = list(dtype) if dtype else _SUPPORTED_TYPES.keys()
-        backend = list(backend) if backend else list(GPUBackend)
-        k_type = list(k_type) if k_type else list(GPUKernel)
-
-        for d, b, k in itertools.product(dtype, backend, k_type):
-            _populate_kernel_cache(d, b, k)
 
 
 def _lombscargle(x, y, freqs, pgram, y_dot, cp_stream, use_numba):
