+++ python/cusignal/_spectral_kernels.py
@@ -1,0 +1,180 @@
+# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# import cupy as cp
+
+from math import sin, cos, atan2
+from numba import cuda, void
+from string import Template
+
+
+def _numba_lombscargle(x, y, freqs, pgram, y_dot):
+    """
+    _lombscargle(x, y, freqs)
+    Computes the Lomb-Scargle periodogram.
+    Parameters
+    ----------
+    x : array_like
+        Sample times.
+    y : array_like
+        Measurement values (must be registered so the mean is zero).
+    freqs : array_like
+        Angular frequencies for output periodogram.
+    Returns
+    -------
+    pgram : array_like
+        Lomb-Scargle periodogram.
+    Raises
+    ------
+    ValueError
+        If the input arrays `x` and `y` do not have the same shape.
+    See also
+    --------
+    lombscargle
+    """
+
+    F = cuda.grid(1)
+    strideF = cuda.gridsize(1)
+
+    if not y_dot[0]:
+        yD = 1.0
+    else:
+        yD = 2.0 / y_dot[0]
+
+    for i in range(F, freqs.shape[0], strideF):
+
+        # Copy data to registers
+        freq = freqs[i]
+
+        xc = 0.0
+        xs = 0.0
+        cc = 0.0
+        ss = 0.0
+        cs = 0.0
+
+        for j in range(x.shape[0]):
+
+            c = cos(freq * x[j])
+            s = sin(freq * x[j])
+
+            xc += y[j] * c
+            xs += y[j] * s
+            cc += c * c
+            ss += s * s
+            cs += c * s
+
+        tau = atan2(2.0 * cs, cc - ss) / (2.0 * freq)
+        c_tau = cos(freq * tau)
+        s_tau = sin(freq * tau)
+        c_tau2 = c_tau * c_tau
+        s_tau2 = s_tau * s_tau
+        cs_tau = 2.0 * c_tau * s_tau
+
+        pgram[i] = (
+            0.5
+            * (
+                (
+                    (c_tau * xc + s_tau * xs) ** 2
+                    / (c_tau2 * cc + cs_tau * cs + s_tau2 * ss)
+                )
+                + (
+                    (c_tau * xs - s_tau * xc) ** 2
+                    / (c_tau2 * ss - cs_tau * cs + s_tau2 * cc)
+                )
+            )
+        ) * yD
+
+
+def _numba_lombscargle_signature(ty):
+    return void(
+        ty[:], ty[:], ty[:], ty[:], ty[:],  # x  # y  # freqs  # pgram  # y_dot
+    )
+
+
+# Custom Cupy raw kernel implementing lombscargle operation
+# Matthew Nicely - mnicely@nvidia.com
+_cupy_lombscargle_src = Template(
+    """
+$header
+
+extern "C" {
+    __global__ void _cupy_lombscargle(
+            const int x_shape,
+            const int freqs_shape,
+            const ${datatype} * __restrict__ x,
+            const ${datatype} * __restrict__ y,
+            const ${datatype} * __restrict__ freqs,
+            ${datatype} * __restrict__ pgram,
+            const ${datatype} * __restrict__ y_dot
+            ) {
+
+        const int tx {
+            static_cast<int>( blockIdx.x * blockDim.x + threadIdx.x ) };
+        const int stride { static_cast<int>( blockDim.x * gridDim.x ) };
+
+        ${datatype} yD {};
+        if ( y_dot[0] == 0 ) {
+            yD = 1.0f;
+        } else {
+            yD = 2.0f / y_dot[0];
+        }
+
+        for ( int tid = tx; tid < freqs_shape; tid += stride ) {
+
+            ${datatype} freq { freqs[tid] };
+
+            ${datatype} xc {};
+            ${datatype} xs {};
+            ${datatype} cc {};
+            ${datatype} ss {};
+            ${datatype} cs {};
+            ${datatype} c {};
+            ${datatype} s {};
+
+            for ( int j = 0; j < x_shape; j++ ) {
+                c = cos( freq * x[j] );
+                s = sin( freq * x[j] );
+
+                xc += y[j] * c;
+                xs += y[j] * s;
+                cc += c * c;
+                ss += s * s;
+                cs += c * s;
+            }
+
+            ${datatype} tau { atan2( 2.0f * cs, cc - ss ) / ( 2.0f * freq ) };
+            ${datatype} c_tau { cos(freq * tau) };
+            ${datatype} s_tau { sin(freq * tau) };
+            ${datatype} c_tau2 { c_tau * c_tau };
+            ${datatype} s_tau2 { s_tau * s_tau };
+            ${datatype} cs_tau { 2.0f * c_tau * s_tau };
+
+            pgram[tid] = (
+                0.5f * (
+                   (
+                       ( c_tau * xc + s_tau * xs )
+                       * ( c_tau * xc + s_tau * xs )
+                       / ( c_tau2 * cc + cs_tau * cs + s_tau2 * ss )
+                    )
+                   + (
+                       ( c_tau * xs - s_tau * xc )
+                       * ( c_tau * xs - s_tau * xc )
+                       / ( c_tau2 * ss - cs_tau * cs + s_tau2 * cc )
+                    )
+                )
+            ) * yD;
+        }
+    }
+}
+"""
+)
